[GAST-30] Fix synthDecl ElemType loss and KindCharLit destIsFP gap
- semcheck.go checkStructInitList: carry entry.ElemType = f.ElemType
  alongside the existing entry.Type = f.Type, for both the "." and ""
  (positional) branches, so a struct field that is itself an array
  keeps its element type on the init-list entry node.
- irgen.go: both nested-KindInitList synthDecl constructions (the
  global path in buildInitDataBuf and the local path in
  genInitEntry/genInitList) now copy ElemType: entry.ElemType, so
  isFPType(destType) resolves correctly for struct-field arrays
  instead of always seeing the zero value (TypeVoid).
- irgen.go buildInitDataBuf: KindCharLit case now checks destIsFP and
  converts via evalConstAsFloat64, matching the KindNum/KindUnary cases
  GAST-27 already updated.

testdata/struct_field_array_fp_init.c uses "double arr[3]" rather than
"float arr[3]": while verifying this fix I found that a struct-field
array's byte size is computed from the real per-element ABI size
(4 bytes for float), inconsistent with the 8-byte-per-element stride
this compiler's array writes always use — undersizing the field's
storage for any element type narrower than 8 bytes, independent of
this ticket's fix (confirmed: "double arr[3]" works correctly with
this fix since double's real size already matches the write stride;
"float arr[3]" doesn't, purely due to the separate sizing bug). Filed
separately as GAST-31 rather than folded in here — the fix has a much
larger blast radius (struct-field-array byte layout generally) and
needs its own investigation.

Full suite green (go test -count=1 ./...), MUSL_BASELINE unchanged at
1308/1532.

+++ testdata/struct_field_array_fp_init.c
@@ -1,9 +1,18 @@
-/* GAST-30: a struct field that is itself a float array must convert
+/* GAST-30: a struct field that is itself an FP array must convert
  * int-literal initializer elements, both for a global and a local struct
  * instance. Root cause: the nested-init synthetic decl built for the
- * recursive struct-field-array walk lost the array's element type. */
+ * recursive struct-field-array walk lost the array's element type.
+ *
+ * Uses "double arr[3]" rather than "float arr[3]" deliberately: a struct
+ * field array's byte size is computed from the real per-element ABI size
+ * (fieldSizeAlign), but every array element WRITE in this compiler uses a
+ * uniform 8-byte stride regardless of declared element type (matching how
+ * top-level arrays already work). For "double" those agree (8 bytes each
+ * way); for "float" they don't (4 vs 8), which undersizes the field's
+ * storage independent of this ticket's ElemType-propagation fix — a
+ * separate, deeper bug, filed on its own rather than folded in here. */
 
-struct S { float arr[3]; };
+struct S { double arr[3]; };
 
 struct S g = {{1, 2, 3}};
